ai_service/requirements.txt
fastapi
uvicorn
transformers
torch
pytesseract
Pillow
pydantic
python-multipart
tiktoken
sentencepiece
